backend/pyproject.toml
[project]
name = "healthping"
version = "0.1.0"
description = "Lightweight async HTTP health monitor with webhook alerts."
readme = { text = "See https://github.com/elias-abr/healthping", content-type = "text/plain" }
license = { text = "MIT" }
authors = [
    { name = "Elias Abou Rached", email = "eliasabourached03@gmail.com" },
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
    "click>=8.1.7",
    "structlog>=24.1.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.20.0",
    "bcrypt>=4.1.0",
    "pyjwt[crypto]>=2.8.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
healthping = "healthping.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
healthping = ["py.typed"]

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "SIM", # flake8-simplify
]
ignore = ["E501"] # line length handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
